Compile without warning with iOS 7.1 SDK

+++ ABFramework/Multi/Model/ABSection.m
@@ -26,7 +26,7 @@
 #pragma mark - Accessors
 -(NSString*) description
 {
-    return [NSString stringWithFormat:@"ABSection: %@ (%i)", self.headerTitle, self.items.count];
+    return [NSString stringWithFormat:@"ABSection: %@ (%i)", self.headerTitle, (int)self.items.count];
 }
 
 @end
